Migrate app from JavaScript to TypeScript for improved type safety and maintainability
